// frontend/types/supabase_new.ts
export type Database = {
  public: {
    Tables: {
// ...
      courses: {
        Row: {
          codes: string[] | null
          credits: number
          description: string | null
          distributions: string[] | null
          flags: string[] | null
          id: string
          is_colsem: boolean | null
          is_fysem: boolean | null
          professors: string[] | null
          requirements: string | null
          seasons: string[] | null
          term: string
          title: string
        }
        Insert: {
          codes?: string[] | null
          credits: number
          description?: string | null
          distributions?: string[] | null
          flags?: string[] | null
          id: string
          is_colsem?: boolean | null
          is_fysem?: boolean | null
          professors?: string[] | null
          requirements?: string | null
          seasons?: string[] | null
          term: string
          title: string
        }
        Update: {
          codes?: string[] | null
          credits?: number
          description?: string | null
          distributions?: string[] | null
          flags?: string[] | null
          id?: string
          is_colsem?: boolean | null
          is_fysem?: boolean | null
          professors?: string[] | null
          requirements?: string | null
          seasons?: string[] | null
          term?: string
          title?: string
        }
        Relationships: []
      }
// ...
      fyp: {
        Row: {
          id: number
          language_placement: string | null
          name: string | null
          term_arrangement: string | null
          user_id: string | null
        }
        Insert: {
          id?: number
          language_placement?: string | null
          name?: string | null
          term_arrangement?: string | null
          user_id?: string | null
        }
        Update: {
          id?: number
          language_placement?: string | null
          name?: string | null
          term_arrangement?: string | null
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "fyp_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      student_courses: {
        Row: {
          course_id: string | null
          fyp_id: number | null
          id: number
          result: string
          status: string
          term: string
        }
        Insert: {
          course_id?: string | null
          fyp_id?: number | null
          id?: number
          result: string
          status: string
          term: string
        }
        Update: {
          course_id?: string | null
          fyp_id?: number | null
          id?: number
          result?: string
          status?: string
          term?: string
        }
        Relationships: [
          {
            foreignKeyName: "student_courses_course_id_fkey"
            columns: ["course_id"]
            isOneToOne: false
            referencedRelation: "courses"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "student_courses_fyp_id_fkey"
            columns: ["fyp_id"]
            isOneToOne: false
            referencedRelation: "fyp"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      users: {
        Row: {
          id: string
          name: string | null
          net_id: string
        }
        Insert: {
          id: string
          name?: string | null
          net_id: string
        }
        Update: {
          id?: string
          name?: string | null
          net_id?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never